requirements.txt
aiofiles==24.1.0
aiogram==3.18.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.6.3
attrs==24.2.0
cachetools==4.2.2
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==43.0.1
distro==1.9.0
frozenlist==1.4.1
greenlet==3.1.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.8
jiter==0.6.1
magic-filter==1.0.12
multidict==6.1.0
openai==1.51.2
packaging==24.1
pdf2image==1.17.0
pdfminer.six==20240706
pillow==10.4.0
propcache==0.2.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyPDF2==3.0.1
pytesseract==0.3.13
python-dotenv==1.0.1
python-telegram-bot==20.3
pytz==2024.1
requests==2.32.3
schedule==1.2.2
setuptools==74.1.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.34
tornado==6.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.15
yarl==1.14.0
